+++ src/index.ts
@@ -7,5 +7,8 @@ export * from "./tasks-queue.module.js";
 export * from "./tasks-queue.service.js";
 export * from "./tasks-queue-async-options.js";
 export * from "./tasks-worker.js";
+export * from "./multi-step-payload.js";
+export * from "./multi-step-task.js";
+export * from "./sequential-task.js";
 export * from "./manage.model.js";
 export * from "./manage-tasks-queue.service.js";